backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
websockets>=14.0
pydantic>=2.0
gitpython>=3.1.0
anthropic>=0.52.0

# Testing
pytest>=8.0
pytest-asyncio>=0.25.0
httpx>=0.28.0
